+++ src/ghreq/__init__.py
@@ -112,6 +112,7 @@ class Client:
         token: str | None = None,
         api_url: str = DEFAULT_API_URL,
         session: requests.Session | None = None,
+        set_headers: bool | None = None,
         user_agent: str | None = None,
         accept: str | None = DEFAULT_ACCEPT,
         api_version: str | None = DEFAULT_API_VERSION,
@@ -123,18 +124,24 @@ class Client:
         :param token:
             The GitHub access token, if any, to use to authenticate to the API.
 
-            This argument is ignored if a non-`None` ``session`` is supplied.
+            This argument is ignored if ``set_headers`` is `False` or defaulted
+            to `False`.
 
         :param api_url:
             The base URL to which to append paths passed to the request methods
 
         :param session:
             A pre-configured `requests.Session` instance to use for making
-            requests.
+            requests.  If no session is supplied, a new session is
+            instantiated.
 
-            If no session is supplied, `Client` instantiates a new session and
-            sets the following request headers on it.  (These headers are not
-            set on sessions passed to the constructor.)
+        :param set_headers:
+            Whether to set various headers for requests made via the session.
+            If ``set_headers`` is `None`, it is defaulted to `True` if
+            ``session`` is `None` and to `False` otherwise.
+
+            If ``set_headers`` is `True` or defaulted to `True`, the following
+            request headers are set on the session:
 
             - :mailheader:`Accept` (if ``accept`` is non-`None`)
             - :mailheader:`Authorization` (set to ``"Bearer {token}"`` if
@@ -144,29 +151,37 @@ class Client:
               non-`None`)
             - any additional headers included in ``headers``
 
+            If ``set_headers`` is `False` or defaulted to `False`, then
+            `Client` does not set any headers on the session, and the other
+            header-related parameters are ignored.
+
         :param user_agent:
             A user agent string to include in the headers of requests.  If not
             set, the `requests` library's default user agent is used.
 
-            This argument is ignored if a non-`None` ``session`` is supplied.
+            This argument is ignored if ``set_headers`` is `False` or defaulted
+            to `False`.
 
         :param accept:
             Value to set the :mailheader:`Accept` header to.  Can be set to
             `None` to not set the header at all.
 
-            This argument is ignored if a non-`None` ``session`` is supplied.
+            This argument is ignored if ``set_headers`` is `False` or defaulted
+            to `False`.
 
         :param api_version:
             Value to set the :mailheader:`X-GitHub-Api-Version` header to.  Can
             be set to `None` to not set the header at all.
 
-            This argument is ignored if a non-`None` ``session`` is supplied.
+            This argument is ignored if ``set_headers`` is `False` or defaulted
+            to `False`.
 
         :param headers:
             Optional mapping of additional headers to set on the session after
             setting all other headers.
 
-            This argument is ignored if a non-`None` ``session`` is supplied.
+            This argument is ignored if ``set_headers`` is `False` or defaulted
+            to `False`.
 
         :param mutation_delay:
             When making a ``POST``, ``PATCH``, ``PUT``, or ``DELETE`` request,
@@ -183,6 +198,11 @@ class Client:
         self.api_url = api_url
         if session is None:
             session = requests.Session()
+            if set_headers is None:
+                set_headers = True
+        elif set_headers is None:
+            set_headers = False
+        if set_headers:
             if token is not None:
                 session.headers["Authorization"] = f"Bearer {token}"
             if user_agent is not None:
@@ -193,7 +213,6 @@ class Client:
                 session.headers["Accept"] = accept
             if headers is not None:
                 session.headers.update(headers)
-        # No headers are set on pre-supplied sessions
         self.session = session
         # GitHub recommends waiting 1 second between non-GET requests in order
         # to avoid hitting secondary rate limits.